разбивает модуль со вспомогательными функциями на два: общий и утилиты для задач
