+++ src/view/point-edit-view.js
@@ -1,6 +1,6 @@
 import { DEFAULT_POINT, TYPES } from '../mock/const.js';
 import AbstractView from '../framework/view/abstract-view.js';
-import { humanizePointDate, DATETIME_FORMAT_NEW } from '../util.js';
+import { humanizePointDate, DATETIME_FORMAT_NEW } from '../util/point.js';
 
 function createPointEditTemplate(point, destinations, offers) {
